For the first challenge, i assigned a value of 0 to "sum" and increment it by the previous value, then i changed return value to sum.
In challenge 2, i assigned count to the value 0, then use modulo character (%) to find remainder of each value in the array. To find even numbers "% 2" is used, and for loop goes over each number in the array and uses the modulo 2 operator on each number to find the even number.

Finally, i change the return, to " return count".